requirements.txt
Django==1.5.2
Pillow==2.1.0
South==0.8.2
django-bootstrap-form==2.0.6
sorl-thumbnail==11.12
psycopg2==2.5.1
django-browserid==0.8
fancy-tag==0.2.0
requests==1.2.3
Jinja2==2.7.1
jingo==0.6.1
MarkupSafe==0.18
geopy==0.95.1
